backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.29
psycopg2-binary>=2.9.9
pydantic>=2.10.0
pydantic-settings>=2.2.1
python-multipart>=0.0.9
aiofiles>=23.2.1
celery>=5.3.6
redis>=5.0.3
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
httpx>=0.27.0

# OCR (optionnel - décommentez si besoin d'extraction PDF/image)
pdfplumber>=0.11.0
pypdf>=5.5.0
pytesseract>=0.3.10
Pillow>=11.0.0
